# public/api/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "POS API",
        "description": "REST API untuk Sistem Point of Sale (POS)",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8303/api",
            "description": "Development Server"
        }
    ],
    "paths": {
...
        "/barangs": {
            "get": {
                "tags": [
                    "Barangs"
                ],
                "summary": "Dapatkan semua barang (pagination)",
                "operationId": "getBarangs",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 15
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "beras"
                        }
                    },
                    {
                        "name": "kategori_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Pagination"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Barangs"
                ],
                "summary": "Buat barang baru",
                "operationId": "createBarang",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BarangCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/barangs/{kode_barang}": {
            "get": {
                "tags": [
                    "Barangs"
                ],
                "summary": "Dapatkan detail barang",
                "operationId": "getBarangDetail",
                "parameters": [
                    {
                        "name": "kode_barang",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "BRG001"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Barangs"
                ],
                "summary": "Update barang",
                "operationId": "updateBarang",
                "parameters": [
                    {
                        "name": "kode_barang",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "BRG001"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "kategori_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "nama": {
                                        "type": "string",
                                        "example": "Beras 10kg"
                                    },
                                    "harga_beli": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 60000
                                    },
                                    "harga_jual": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 75000
                                    },
                                    "stok": {
                                        "type": "integer",
                                        "example": 20
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Barangs"
                ],
                "summary": "Hapus barang",
                "operationId": "deleteBarang",
                "parameters": [
                    {
                        "name": "kode_barang",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "BRG001"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "BarangCreate": {
                "required": [
                    "kategori_id",
                    "nama",
                    "harga_beli",
                    "harga_jual",
                    "stok"
                ],
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/Pagination"
                    },
                    {
                        "properties": {
                            "kode_barang": {
                                "description": "Optional. Jika tidak diisi, sistem akan mengenerate otomatis (prefix BRG + 3 angka).",
                                "type": "string",
                                "example": "BRG123"
                            },
                            "kategori_id": {
                                "type": "integer",
                                "example": 1
                            },
                            "nama": {
                                "type": "string",
                                "example": "Beras 5kg"
                            },
                            "harga_beli": {
                                "type": "number",
                                "format": "float",
                                "example": 50000
                            },
                            "harga_jual": {
                                "type": "number",
                                "format": "float",
                                "example": 65000
                            },
                            "stok": {
                                "type": "integer",
                                "example": 10
                            }
                        },
                        "type": "object"
                    }
                ]
            },
...
            "ApiResponse": {
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "Pagination": {
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    },
                    "first_page_url": {
                        "type": "string"
                    },
                    "from": {
                        "type": "integer",
                        "nullable": true
                    },
                    "last_page": {
                        "type": "integer"
                    },
                    "last_page_url": {
                        "type": "string"
                    },
                    "next_page_url": {
                        "type": "string",
                        "nullable": true
                    },
                    "path": {
                        "type": "string"
                    },
                    "per_page": {
                        "type": "integer"
                    },
                    "prev_page_url": {
                        "type": "string",
                        "nullable": true
                    },
                    "to": {
                        "type": "integer",
                        "nullable": true
                    },
                    "total": {
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}